# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de la plataforma de eventos IA Week",
    "description": "API multi-organización para el registro y la gestión de eventos. La organización se resuelve por el host de la petición.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/events/{event_id}/registrations": {
      "get": {
        "tags": [
          "inscripciones"
        ],
        "summary": "Listar las inscripciones de un evento",
        "operationId": "list_registrations_api_v1_events__event_id__registrations_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending_verification",
                    "pending_approval",
                    "pending_payment",
                    "confirmed",
                    "rejected",
                    "cancelled",
                    "waitlisted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Elementos por página",
              "default": 20,
              "title": "Limit"
            },
            "description": "Elementos por página"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Elementos omitidos",
              "default": 0,
              "title": "Offset"
            },
            "description": "Elementos omitidos"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RegistrationListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{event_id}/registration-questions": {
      "get": {
        "tags": [
          "inscripciones"
        ],
        "summary": "Listar las preguntas personalizadas de inscripción de un evento",
        "operationId": "list_registration_questions_api_v1_events__event_id__registration_questions_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationQuestionResponse"
                  },
                  "title": "Response List Registration Questions Api V1 Events  Event Id  Registration Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inscripciones"
        ],
        "summary": "Añadir una pregunta personalizada de inscripción",
        "operationId": "create_registration_question_api_v1_events__event_id__registration_questions_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationQuestionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationQuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/events/{slug}/registration-questions": {
      "get": {
        "tags": [
          "público"
        ],
        "summary": "Listar las preguntas personalizadas de inscripción de un evento",
        "operationId": "list_registration_questions_api_v1_public_events__slug__registration_questions_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegistrationQuestionPublic"
                  },
                  "title": "Response List Registration Questions Api V1 Public Events  Slug  Registration Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/events/{slug}/registrations": {
      "post": {
        "tags": [
          "público"
        ],
        "summary": "Inscribirse a un evento",
        "description": "Da de alta una inscripción, o reencola el correo pendiente si el email ya estaba inscrito. Responde siempre igual, para no filtrar si un correo ya está inscrito.",
        "operationId": "create_registration_api_v1_public_events__slug__registrations_post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/registrations/verify": {
      "post": {
        "tags": [
          "público"
        ],
        "summary": "Verificar el correo de una inscripción",
        "description": "Consume el token del enlace de verificación y evalúa el estado siguiente.",
        "operationId": "verify_registration_api_v1_public_registrations_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyRegistrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/registrations/confirm-waitlist-promotion": {
      "post": {
        "tags": [
          "público"
        ],
        "summary": "Confirmar una promoción desde la lista de espera",
        "description": "Consume el token del enlace de promoción. Si ha caducado, responde que caducó sin más detalle y deja que el barrido cron reasigne la plaza.",
        "operationId": "confirm_waitlist_promotion_api_v1_public_registrations_confirm_waitlist_promotion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmWaitlistPromotionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmWaitlistPromotionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/registrations/cancel": {
      "post": {
        "tags": [
          "público"
        ],
        "summary": "Cancelar una inscripción por autocancelación",
        "description": "Consume el token del enlace de cancelación (un solo uso). Si liberaba una plaza confirmada, promueve automáticamente a la lista de espera.",
        "operationId": "cancel_registration_api_v1_public_registrations_cancel_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelRegistrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelRegistrationRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "CancelRegistrationRequest",
        "description": "Token del enlace de autocancelación recibido por correo."
      },
      "CancelRegistrationResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "CancelRegistrationResponse",
        "description": "Resultado de cancelar una inscripción por autocancelación."
      },
...
      "ConfirmWaitlistPromotionRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "ConfirmWaitlistPromotionRequest",
        "description": "Token del enlace de confirmación de una promoción de lista de espera."
      },
      "ConfirmWaitlistPromotionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ConfirmWaitlistPromotionResponse",
        "description": "Resultado de confirmar una promoción de lista de espera."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_RegistrationListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RegistrationListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RegistrationListItem]"
      },
...
      "RegistrationAnswerInput": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "question_id"
        ],
        "title": "RegistrationAnswerInput",
        "description": "Respuesta a una pregunta personalizada del formulario de inscripción."
      },
...
      "RegistrationListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_verification",
              "pending_approval",
              "pending_payment",
              "confirmed",
              "rejected",
              "cancelled",
              "waitlisted"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "waitlist_promoted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waitlist Promoted At"
          },
          "waitlist_promotion_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waitlist Promotion Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "status",
          "created_at",
          "verified_at",
          "confirmed_at",
          "waitlist_promoted_at",
          "waitlist_promotion_expires_at"
        ],
        "title": "RegistrationListItem",
        "description": "Fila del listado de inscripciones del panel de organizador."
      },
      "RegistrationMessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "RegistrationMessageResponse",
        "description": "Respuesta genérica del alta: siempre el mismo mensaje, exista o no ya el email."
      },
      "RegistrationQuestionCreate": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "short_text",
              "single_choice",
              "multiple_choice"
            ],
            "title": "Type"
          },
          "label": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "type",
          "label"
        ],
        "title": "RegistrationQuestionCreate",
        "description": "Alta de una pregunta personalizada desde el panel de organizador."
      },
      "RegistrationQuestionPublic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "label",
          "required",
          "sort_order"
        ],
        "title": "RegistrationQuestionPublic",
        "description": "Pregunta personalizada tal como la ve el formulario público."
      },
      "RegistrationQuestionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "short_text",
              "single_choice",
              "multiple_choice"
            ],
            "title": "Type"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "label",
          "required",
          "sort_order",
          "options"
        ],
        "title": "RegistrationQuestionResponse",
        "description": "Pregunta personalizada tal como la ve el panel de organizador."
      },
...
      "SubmitRegistrationRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/RegistrationAnswerInput"
            },
            "type": "array",
            "title": "Answers"
          },
          "data_processing_accepted": {
            "type": "boolean",
            "title": "Data Processing Accepted",
            "description": "Consentimiento de tratamiento de datos, obligatorio para inscribirse."
          },
          "marketing_accepted": {
            "type": "boolean",
            "title": "Marketing Accepted",
            "default": false
          },
          "recording_accepted": {
            "type": "boolean",
            "title": "Recording Accepted",
            "default": false
          },
          "turnstile_token": {
            "type": "string",
            "title": "Turnstile Token",
            "description": "Token del widget de Turnstile"
          }
        },
        "type": "object",
        "required": [
          "email",
          "full_name",
          "data_processing_accepted",
          "turnstile_token"
        ],
        "title": "SubmitRegistrationRequest",
        "description": "Datos del formulario público de inscripción a un evento."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifyRegistrationRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "VerifyRegistrationRequest",
        "description": "Token del enlace de verificación recibido por correo."
      },
      "VerifyRegistrationResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "message",
          "status"
        ],
        "title": "VerifyRegistrationResponse",
        "description": "Resultado de verificar una inscripción."
      }
    }
  }
}